backend/requirements.txt
fastapi
uvicorn
chromadb
openai
pydantic
python-jose
passlib[bcrypt]
pymongo
networkx
python-multipart 